Casting some time_t's to long's to avoid compiler warnings.

+++ src/util.c
@@ -1,5 +1,5 @@
 /* -*- c-basic-offset: 4; indent-tabs-mode: nil -*-
- * $Id: util.c,v 1.178 2005/08/18 15:27:55 we7u Exp $
+ * $Id: util.c,v 1.179 2005/08/23 19:36:42 we7u Exp $
  *
  * XASTIR, Amateur Station Tracking and Information Reporting
  * Copyright (C) 1999,2000  Frank Giannandrea
@@ -1402,17 +1402,17 @@ time_t time_from_aprsstring(char *aprs_time) {
             alert_time.tm_min);
 
         a_time = mktime(&alert_time);
-        fprintf(stderr,"Alert: %ld\n", a_time);
+        fprintf(stderr,"Alert: %ld\n", (long)a_time);
 
         now_time = sec_now();
-        fprintf(stderr,"  Now: %ld\n", now_time);
+        fprintf(stderr,"  Now: %ld\n", (long)now_time);
 
         diff = now_time - a_time;
 
         if (diff >= 0)
-            fprintf(stderr,"Expired by %ld minutes\n", diff/60);
+            fprintf(stderr,"Expired by %ld minutes\n", (long)(diff/60) );
         else
-            fprintf(stderr,"%ld minutes until expiration\n", (-diff)/60);
+            fprintf(stderr,"%ld minutes until expiration\n", (long)((-diff)/60) );
 
         if (alert_time.tm_isdst > 0)
             fprintf(stderr,"Daylight savings time is in effect\n");
